채팅완료(ai챗이랑 디자인 병합X)

+++ src/main/java/com/boot/trade/dao/TradePostDAO.java
@@ -40,5 +40,5 @@ public interface TradePostDAO {
 	int getChatCountByPostID(int postID);
 	
 	// 채팅용
-    TradePostDTO getTradePostById(int postID);
+    TradePostDTO getTradePostById(Long postID);
 }
